+++ handlers/go.mod
@@ -1,0 +1,13 @@
+module api/internal/handlers
+
+go 1.21.3
+
+require github.com/go-rod/rod v0.114.5
+
+require (
+	github.com/ysmood/fetchup v0.2.3 // indirect
+	github.com/ysmood/goob v0.4.0 // indirect
+	github.com/ysmood/got v0.34.1 // indirect
+	github.com/ysmood/gson v0.7.3 // indirect
+	github.com/ysmood/leakless v0.8.0 // indirect
+)
