+++ src/go.mod
@@ -1,0 +1,8 @@
+module api/internal/src
+
+go 1.21.3
+
+require (
+	github.com/go-chi/chi v1.5.5
+	github.com/lib/pq v1.10.9
+)
